// src/models/interfaces/product.interface.ts
// ! Imports
// * Services
import mongoose from '../../services/mongoose.service';

// ! Interfaces Defintions
interface idsInArrayMethodInterface {
	state: boolean;
	missingProductIds: mongoose.Types.ObjectId[];
	withoutStock: mongoose.Types.ObjectId[];
}
interface genericProductPropertiesInterface {
	_id: mongoose.Types.ObjectId;
	categoryId: number;
	name: string;
	description: string;
	price: number;
	imagesURL: string[];
	timeStamp: Date;
	memory: number;
}
interface productPropertiesInterface extends genericProductPropertiesInterface {
	stock: number[];
	colors: string[];
}
interface cartProductsInterface extends genericProductPropertiesInterface {
	color: string;
	amount: number;
}

// ! Exports
export {
	productPropertiesInterface,
	idsInArrayMethodInterface,
	genericProductPropertiesInterface,
	cartProductsInterface,
};